feat(cloud-core): shared web/desktop cloud domain — Phase 1 (ISSUE-022)
The web frontend and the desktop cloud panel each fully re-implemented the same
cloud functionality (scopes, connectors, MCP/sandbox endpoints, repo identity,
provider registry) against the same backend, differing only in auth/transport
(Supabase token vs Electron IPC). This extracts the shared, transport-agnostic
domain into @puppyone/cloud-core and makes the web consume it — zero behaviour
change, single source of truth.

- packages/cloud-core: platform-agnostic domain (types + API-client factories +
  pure helpers) with a CloudTransport seam for per-platform auth/HTTP. Covers
  scopes, connectors + repo identity, MCP endpoints, sandbox endpoints, and the
  access-provider registry. No Next.js/Electron/Supabase/SWR imports (enforced
  by check-data-ui-boundaries).
- frontend: lib/cloudCoreTransport.ts binds cloud-core to apiClient (a straight
  pass-through). lib/{repoApi,mcpEndpointsApi,sandboxEndpointsApi,
  accessProviderRegistry}.ts are now thin bindings/re-exports — call sites
  unchanged. tsconfig path alias wires it (Next honours tsconfig paths).
- scripts/check-data-ui-boundaries.mjs: enforce cloud-core stays platform-agnostic.

Verified: cd frontend && tsc --noEmit → 0 errors (all repoApi/registry call
sites still resolve); boundary check passes. See packages/cloud-core/ROADMAP.md
for Phases 2-4 (desktop submodule consumption, remaining clients/hooks, UI).
